pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markerseek"
version = "0.1.0"
description = "Nucleotide diversity analysis for chloroplast GenBank collections."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {text = "MIT"}
authors = [
  {name = "Gao Chengwen"}
]
dependencies = [
  "biopython>=1.84",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
markerseek = "markerseek.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
